Move ProposedCircuit and builder into own file
 Moves ProposedCircuit into
 store::proposed_circuit::{ProposedCircuit, ProposedCircuitBuilder}

 Structs are still exposed only in store:: via pub use. This change
 also requires getters to be implemented on the structs to be used
 throughout the store implementations.
