+++ libsplinter/src/admin/store/builders.rs
@@ -17,249 +17,7 @@
 use crate::admin::messages::is_valid_circuit_id;
 
 use super::error::BuilderError;
-use super::{
-    AuthorizationType, CircuitProposal, DurabilityType, PersistenceType, ProposalType,
-    ProposedCircuit, ProposedNode, ProposedService, RouteType, VoteRecord,
-};
-
-/// Builder to be used to build a `ProposedCircuit` which will be included in a `CircuitProposal`
-#[derive(Default, Clone)]
-pub struct ProposedCircuitBuilder {
-    circuit_id: Option<String>,
-    roster: Option<Vec<ProposedService>>,
-    members: Option<Vec<ProposedNode>>,
-    authorization_type: Option<AuthorizationType>,
-    persistence: Option<PersistenceType>,
-    durability: Option<DurabilityType>,
-    routes: Option<RouteType>,
-    circuit_management_type: Option<String>,
-    application_metadata: Option<Vec<u8>>,
-    comments: Option<String>,
-}
-
-impl ProposedCircuitBuilder {
-    /// Creates a new proposed circuit builder
-    pub fn new() -> Self {
-        ProposedCircuitBuilder::default()
-    }
-
-    // Returns the circuit ID in the builder
-    pub fn circuit_id(&self) -> Option<String> {
-        self.circuit_id.clone()
-    }
-
-    /// Returns the list of services in the builder
-    pub fn roster(&self) -> Option<Vec<ProposedService>> {
-        self.roster.clone()
-    }
-
-    /// Returns the list of node IDs in the builder
-    pub fn members(&self) -> Option<Vec<ProposedNode>> {
-        self.members.clone()
-    }
-
-    /// Returns the authorization type in the builder
-    pub fn authorization_type(&self) -> Option<AuthorizationType> {
-        self.authorization_type.clone()
-    }
-
-    /// Returns the persistence type in the builder
-    pub fn persistence(&self) -> Option<PersistenceType> {
-        self.persistence.clone()
-    }
-
-    /// Returns the durability type in the builder
-    pub fn durability(&self) -> Option<DurabilityType> {
-        self.durability.clone()
-    }
-
-    /// Returns the routing type in the builder
-    pub fn routes(&self) -> Option<RouteType> {
-        self.routes.clone()
-    }
-
-    /// Returns the circuit management type in the builder
-    pub fn circuit_management_type(&self) -> Option<String> {
-        self.circuit_management_type.clone()
-    }
-
-    /// Returns the appplication metdata in the builder
-    pub fn application_metadata(&self) -> Option<Vec<u8>> {
-        self.application_metadata.clone()
-    }
-
-    /// Returns the comments describing the circuit proposal in the builder
-    pub fn comments(&self) -> Option<String> {
-        self.comments.clone()
-    }
-
-    /// Sets the circuit ID
-    ///
-    /// # Arguments
-    ///
-    ///  * `circuit_id` - The unique ID of the circuit
-    pub fn with_circuit_id(mut self, circuit_id: &str) -> ProposedCircuitBuilder {
-        self.circuit_id = Some(circuit_id.into());
-        self
-    }
-
-    /// Sets the list of services that are included in the circuit
-    ///
-    /// # Arguments
-    ///
-    ///  * `services` - List of proposed services
-    pub fn with_roster(mut self, services: &[ProposedService]) -> ProposedCircuitBuilder {
-        self.roster = Some(services.into());
-        self
-    }
-
-    /// Sets the list of nodes that are included in the circuit
-    ///
-    /// # Arguments
-    ///
-    ///  * `members` - List of proposed nodes
-    pub fn with_members(mut self, members: &[ProposedNode]) -> ProposedCircuitBuilder {
-        self.members = Some(members.into());
-        self
-    }
-
-    /// Sets the authorization type
-    ///
-    /// # Arguments
-    ///
-    ///  * `auth` - The authorization type for the circuit
-    pub fn with_authorization_type(mut self, auth: &AuthorizationType) -> ProposedCircuitBuilder {
-        self.authorization_type = Some(auth.clone());
-        self
-    }
-
-    /// Sets the persistence type
-    ///
-    /// # Arguments
-    ///
-    ///  * `persistence` - The persistence type for the circuit
-    pub fn with_persistence(mut self, persistence: &PersistenceType) -> ProposedCircuitBuilder {
-        self.persistence = Some(persistence.clone());
-        self
-    }
-
-    /// Sets the durability type
-    ///
-    /// # Arguments
-    ///
-    ///  * `durability` - The durability type for the circuit
-    pub fn with_durability(mut self, durability: &DurabilityType) -> ProposedCircuitBuilder {
-        self.durability = Some(durability.clone());
-        self
-    }
-
-    /// Sets the routes type
-    ///
-    /// # Arguments
-    ///
-    ///  * `routes` - The routes type for the circuit
-    pub fn with_routes(mut self, route_type: &RouteType) -> ProposedCircuitBuilder {
-        self.routes = Some(route_type.clone());
-        self
-    }
-
-    /// Sets the circuit managment type
-    ///
-    /// # Arguments
-    ///
-    ///  * `circuit_management_type` - The circuit_management_type for the circuit
-    pub fn with_circuit_management_type(
-        mut self,
-        circuit_management_type: &str,
-    ) -> ProposedCircuitBuilder {
-        self.circuit_management_type = Some(circuit_management_type.into());
-        self
-    }
-
-    /// Sets the application metadata
-    ///
-    /// # Arguments
-    ///
-    ///  * `application_metadata` - The application_metadata for the proposed circuit
-    pub fn with_application_metadata(
-        mut self,
-        application_metadata: &[u8],
-    ) -> ProposedCircuitBuilder {
-        self.application_metadata = Some(application_metadata.into());
-        self
-    }
-
-    /// Sets the comments
-    ///
-    /// # Arguments
-    ///
-    ///  * `comments` - The comments describing the purpose of the proposed circuit
-    pub fn with_comments(mut self, comments: &str) -> ProposedCircuitBuilder {
-        self.comments = Some(comments.into());
-        self
-    }
-
-    /// Builds a `ProposedCircuit`
-    ///
-    /// Returns an error if the circuit ID, roster, members or circuit management
-    /// type are not set.
-    pub fn build(self) -> Result<ProposedCircuit, BuilderError> {
-        let circuit_id = match self.circuit_id {
-            Some(circuit_id) if is_valid_circuit_id(&circuit_id) => circuit_id,
-            Some(circuit_id) => {
-                return Err(BuilderError::InvalidField(format!(
-                    "circuit_id is invalid ({}): must be an 11 character string composed of two, \
-                     5 character base62 strings joined with a '-' (example: abcDE-F0123)",
-                    circuit_id,
-                )))
-            }
-            None => return Err(BuilderError::MissingField("circuit_id".to_string())),
-        };
-
-        let roster = self
-            .roster
-            .ok_or_else(|| BuilderError::MissingField("roster".to_string()))?;
-
-        let members = self
-            .members
-            .ok_or_else(|| BuilderError::MissingField("members".to_string()))?;
-
-        let authorization_type = self
-            .authorization_type
-            .unwrap_or_else(|| AuthorizationType::Trust);
-
-        let persistence = self.persistence.unwrap_or_else(PersistenceType::default);
-
-        let durability = self
-            .durability
-            .unwrap_or_else(|| DurabilityType::NoDurability);
-
-        let routes = self.routes.unwrap_or_else(RouteType::default);
-
-        let circuit_management_type = self
-            .circuit_management_type
-            .ok_or_else(|| BuilderError::MissingField("circuit_management_type".to_string()))?;
-
-        let application_metadata = self.application_metadata.unwrap_or_default();
-
-        let comments = self.comments.unwrap_or_default();
-
-        let create_circuit_message = ProposedCircuit {
-            circuit_id,
-            roster,
-            members,
-            authorization_type,
-            persistence,
-            durability,
-            routes,
-            circuit_management_type,
-            application_metadata,
-            comments,
-        };
-
-        Ok(create_circuit_message)
-    }
-}
+use super::{CircuitProposal, ProposalType, ProposedCircuit, VoteRecord};
 
 /// Builder to be used to build a `CircuitProposal`
 #[derive(Clone, Default)]
