+++ libsplinter/src/admin/store/diesel/operations/update_proposal.rs
@@ -80,8 +80,7 @@ impl<'a> AdminServiceStoreUpdateProposalOperation
             let proposed_circuit_model = ProposedCircuitModel::from(&proposal.circuit);
             update(proposed_circuit::table.find(&proposal.circuit_id))
                 .set((
-                    proposed_circuit::authorization_type
-                        .eq(proposed_circuit_model.authorization_type),
+                    proposed_circuit::authorization_type.eq(proposed_circuit_model.authorization_type),
                     proposed_circuit::persistence.eq(proposed_circuit_model.persistence),
                     proposed_circuit::durability.eq(proposed_circuit_model.durability),
                     proposed_circuit::routes.eq(proposed_circuit_model.routes),
@@ -260,8 +259,7 @@ impl<'a> AdminServiceStoreUpdateProposalOperation
             let proposed_circuit_model = ProposedCircuitModel::from(&proposal.circuit);
             update(proposed_circuit::table.find(&proposal.circuit_id))
                 .set((
-                    proposed_circuit::authorization_type
-                        .eq(proposed_circuit_model.authorization_type),
+                    proposed_circuit::authorization_type.eq(proposed_circuit_model.authorization_type),
                     proposed_circuit::persistence.eq(proposed_circuit_model.persistence),
                     proposed_circuit::durability.eq(proposed_circuit_model.durability),
                     proposed_circuit::routes.eq(proposed_circuit_model.routes),
